인터페이스 추출
