+++ src/main/java/ladder/creator/LadderCreator.java
@@ -1,0 +1,9 @@
+package ladder.creator;
+
+import ladder.Position;
+import ladder.Row;
+
+public interface LadderCreator {
+    void drawLine(Position row, Position col);
+    Row[] getRows();
+}
