+++ src/main/java/ladder/creator/LadderGameFactory.java
@@ -6,7 +6,7 @@ import ladder.LadderSize;
 public class LadderGameFactory {
 
     public static LadderGame createRandomLadderGame(LadderSize ladderSize) {
-        BasicLadderCreator creator = new BasicLadderCreator(ladderSize);
+        RandomLadderCreator creator = new RandomLadderCreator(ladderSize);
         creator.drawRandomLines(ladderSize);
         return new LadderGame(creator);
     }
